+++ crankshaft-monitor/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+* Added support for the `ImagePull{Started,Failed,Finished}` events ([#](https://github.com/stjude-rust-labs/crankshaft/pull/)).
+
 ## 0.2.0 - 02-11-2026
 
 ### Changed